Add shim ↔ NativeTerm pipe protocol and the real shim
- native-term-win: user SID, logon session id, SDDL, file ACL helpers
  (config's acl module now uses them)
- native-term-session: per-user pipe (owner-only DACL, remote clients
  rejected, single instance, non-blocking duplex), versioned JSON-lines
  protocol, exit classification; accepts clients that already closed
- native-term-shim: runs ssh with the LocalCommand login helper and
  default keepalives, reports lifecycle over a retrying link with state
  replay, reconnect/close prompt, placeholder mode without a host;
  end-to-end tests against a fake ssh
- docs: protocol section, restored-session behavior, roadmap

+++ prototypes/uia-probe/src/main.rs
@@ -45,7 +45,7 @@ fn is_test_terminal(window: &UIElement) -> bool {
     if dir == "*" {
         return true;
     }
-    let pid = window.get_process_id().unwrap_or(0) as u32;
+    let pid = window.get_process_id().unwrap_or(0);
     process_image(pid).is_some_and(|p| p.to_lowercase().starts_with(&format!("{}\\", dir.to_lowercase())))
 }
